Support coreos installer test for s390x
A pxe boot solution for the s390x coreos installer test. It took a bit of tweaks to work,
the main issue being the pxe image is not available in-built but has to be built with
the kernel and initrd. Made the script a bit bulky, but I tried to make it
as clean as possible.